k7_Guru/k7_Guru/requirements.txt
boto3
botocore
dj_database_url
Django
djangorestframework
python-dotenv
djangorestframework
psycopg2-binary
boto3
python-dotenv
django-cors-headers
dj-database-url